requirements-mac.txt
opencv-python
tensorflow-macos
tensorflow-metal
numpy
deep_translator
flask
matplotlib
